src/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
datasets==2.15.0
dill==0.3.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.4
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.9.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.19.4
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.0
orjson==3.9.7
packaging==23.2
pandas==2.1.4
Pillow==10.0.1
protobuf==3.20.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyparsing==3.1.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
thefuzz==0.20.0
threadpoolctl==3.2.0
tinysegmenter==0.3
tldextract==3.6.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.3
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
underthesea==6.8.0
underthesea_core==1.0.4
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
